requirements.txt
certifi==2019.11.28
chardet==3.0.4
cloudinary==1.20.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1
django-allauth==0.41.0
django-cloudinary-storage==0.2.3
django-crispy-forms==1.8.1
django-heroku==0.3.1
django-suit==0.2.28
gunicorn==20.0.4
idna==2.8
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==6.2.1
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-http-client==3.2.4
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
sendgrid==6.1.1
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
urllib3==1.25.7
whitenoise==5.0.1
WTForms==2.2.1
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
